Migrating away from registering for the system directly in the wordpress settings page and offloading to a more secure register page, remove email field setting, remove unregister button, updated readme with new instructions

+++ components/settings.php
@@ -14,22 +14,25 @@ function shift8_cdn_create_menu() {
 
 // Register admin settings
 function register_shift8_cdn_settings() {
-    //register our settings
+    //Register our settings
     register_setting( 'shift8-cdn-settings-group', 'shift8_cdn_enabled' );
     register_setting( 'shift8-cdn-settings-group', 'shift8_cdn_url', 'shift8_cdn_url_validate' );
-    register_setting( 'shift8-cdn-settings-group', 'shift8_cdn_email', 'shift8_cdn_email_validate' );
     register_setting( 'shift8-cdn-settings-group', 'shift8_cdn_api' );
     register_setting( 'shift8-cdn-settings-group', 'shift8_cdn_prefix' );
     register_setting( 'shift8-cdn-settings-group', 'shift8_cdn_css', array( 'default' => 'on' ) );
     register_setting( 'shift8-cdn-settings-group', 'shift8_cdn_js', array( 'default' => 'on' ) );
     register_setting( 'shift8-cdn-settings-group', 'shift8_cdn_media', array( 'default' => 'on' ));
+
+    // Cleanup of old settings no longer needed
+    if (get_option('shift8_cdn_email')) {
+        delete_option('shift8_cdn_email');
+    }
 }
 
 // Uninstall hook
 function handle_shift8_cdn_uninstall_hook() {
   delete_option('shift8_cdn_enabled');
   delete_option('shift8_cdn_url');
-  delete_option('shift8_cdn_email');
   delete_option('shift8_cdn_api');
   delete_option('shift8_cdn_prefix');
   delete_option('shift8_cdn_css');
@@ -52,24 +55,11 @@ function shift8_cdn_url_validate($data){
    	}
 }
 
-function shift8_cdn_email_validate($data){
-	if(filter_var($data, FILTER_SANITIZE_EMAIL)) {
-   		return $data;
-   	} else {
-   		add_settings_error(
-            'shift8_cdn_email',
-            'shift8-cdn-notice',
-            'You did not enter a valid string for the email field',
-            'error');
-   	}
-}
-
 // Validate admin options
 function shift8_cdn_check_enabled() {
   // If enabled is not set
   if(esc_attr( get_option('shift8_cdn_enabled') ) != 'on') return false;
   if(empty(esc_attr(get_option('shift8_cdn_url') ))) return false;
-  if(empty(esc_attr(get_option('shift8_cdn_email') ))) return false;
   if(empty(esc_attr(get_option('shift8_cdn_api') ))) return false;
   if(empty(esc_attr(get_option('shift8_cdn_prefix') ))) return false;
 
@@ -80,7 +70,6 @@ function shift8_cdn_check_enabled() {
 function shift8_cdn_check_options() {
   $shift8_options = array();
   $shift8_options['cdn_url'] = esc_attr( get_option('shift8_cdn_url') );
-  $shift8_options['cdn_email'] = esc_attr( get_option('shift8_cdn_email') );
   $shift8_options['cdn_api'] = esc_attr( get_option('shift8_cdn_api') );
   $shift8_options['cdn_prefix'] = esc_attr( get_option('shift8_cdn_prefix') );
   $shift8_options['static_css'] = esc_attr( get_option('shift8_cdn_css', 'on') );
@@ -89,3 +78,4 @@ function shift8_cdn_check_options() {
 
   return $shift8_options;
 }
+
